moved HomeController and LayoutAdvice into main package  TODO: separate Views from Controllers, ie package edu.msudenver.cs3250.group6.msubanner.controller, .view, etc
