+++ src/main/java/edu/msudenver/cs3250/group6/msubanner/HomeController.java
@@ -1,4 +1,4 @@
-package edu.msudenver.cs3250.group6.views;
+package edu.msudenver.cs3250.group6.msubanner;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 // the purpose here is to get a mapping to the index.html in templates
 // from: https://spring.io/blog/2016/11/21/the-joy-of-mustache-server-side-templates-for-the-jvm
-// TODO: make it work for 
+// TODO: make it work for /index.html too
 @Controller
 class HomeController {
   @GetMapping("/")
