+++ src/main/java/edu/msudenver/cs3250/group6/msubanner/User.java
@@ -1,17 +1,16 @@
 package edu.msudenver.cs3250.group6.msubanner;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
+// import javax.persistence.GeneratedValue;
+// import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Table;
+// import javax.persistence.Table;
 
 /**
  * Persistent User class.
  * @author Group 6
  */
-@Entity
-@Table(name = "users")
+@Entity //@Table(name = "users")
 public class User {
 
     /**
