+++ src/main/resources/templates/index.html
@@ -1,4 +1,5 @@
-{{>header}}
+{{#layout}}
     <h1>Demo</h1>
     <div>Hello World</div>
-{{>footer}}
+{{/layout}}
+{{>layout}}
